pytest-lsp: Move test client protocol into its own file
